nodes/koi-net-coordinator-node/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coordinator-node"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "httpx",
    "python-dotenv",
    "rid-lib>=3.2.1",
    "koi-net==1.0.0b12",
    "rich",
    "ruamel.yaml",
    "python-dotenv"
]

[tool.setuptools]
package-dir = {"" = "."}
